Added command line argument to enable pre-training test
A command line argument '--test_eval' was added to allow the user to determine whether to test the network before training. This provides more flexibility for users in their usage of the script. Pre-testing is off by default, and can be turned on using the argument. 'Network.evaluate' now only runs if '--test_eval' is set to true.