Replace GroupNormalization with BatchNormalization
GroupNormalization layers in ASR_Network.py were replaced with BatchNormalization layers. The modification aims to improve the model's training speed and the quality of the model, as BatchNormalization normalizes the activation of the previous layer at each batch, maintaining the activation means close to 0 and the activation standard deviation close to 1.